refactored configuration manager, added interface
