+++ AppConfig/IConfigManager.cs
@@ -1,0 +1,10 @@
+﻿using UsersServer.Database;
+
+namespace UsersServer.AppConfig
+{
+    public interface IConfigManager
+    {
+        ConnectionStringData GetConnectionString();
+        void SetConnectionString(IConnectionString connectionString);
+    }
+}
